Create osteocyteMetadataCollection
This script asks the user for the data directory in which metadata is stored, and any .czi files for which they would like metadata collected. If the files are of the right format (.czi), are appropriately named (ex: samp5c_regT_sec2_rotY_date20200825.czi), and are not present in the database, they are added and a new file is created.