Create getLacunarCounts.m
Given a binary volume (volume), this function gets all the connected objects. It then output the number objects in the top third of the volume (num_top), middle third of the volume (num_mid), and bottom third of the volume (num_bot). It also outputs a labeled volume (L) with top cells labeled 1, middle cells 2, and bottom cells 3.