Create osteocyteLacunarAnalysis.m
This script analyzes pre-cropped and rotated tif stacks, as output by osteocyteLacunarProcessing.m Given the analysis settings (can change these in the code), the script determines the number of osteocytes in the top, middle, and bottom sections of the volume. Outputs this data in an Excel sheet and a labeled volume.