Create createSqlValueList.m
This function is used by the updated osteocyteLacunarAnalysis.m, osteocyteLacunarProcessing.m, and osteocyteMetadataCollection.m scripts to pull from and insert values into the osteocyte.db SQL database. Given a cell array of values to add to a SQL table (values_to_add), outputs the list of values as a string in the proper format (out).  That is, values are delimited by commas, and strings have double quotes around them.